Pacioli.jl
The user file. Contains program to add accounting entries to dictionaries, to create lists of transactions, and running balances for each account.  Also generates income statements, and balance sheets.

latest changes
returns a in program error, when the user misspells either the name of the ledger, or the name of an account or an account is assigned to the wrong ledger.